Moves graph service files to their own folder
